Fix types, dynamic allocation, and dx value

+++ pointGenerators.c
@@ -5,10 +5,10 @@
 #include "pointGenerators.h"
 
 void xPoints(
-        double dx,
-        double x,
-        double * points,
-        double f(double)
+        float dx,
+        float x,
+        float * points,
+        float f(float)
         ){
 
     for (unsigned int i = 0; i < num; i++){
